fix(doom): Adds RemGauntlet
